+++ app/src/main/java/com/iua/agustinpereyra/controller/CattleManager.kt
@@ -1,5 +1,6 @@
 package com.iua.agustinpereyra.controller
 
+import android.content.Context
 import com.iua.agustinpereyra.repository.database.entities.Cattle
 
 class CattleManager {
@@ -22,6 +23,30 @@ class CattleManager {
                 ORDER_BY_WEIGHT_DESC -> cattleList.sortedByDescending { it.weight }
                 else -> cattleList
             }
+        }
+
+        fun orderBasedOnPreferences(originalList: List<Cattle>, context: Context?): List<Cattle> {
+            var cattleList = originalList
+            val preferenceUtils = PreferenceUtils(context)
+            // Filtering
+            // Check sex filter
+            val sexFilterStatus = preferenceUtils.getSexFilter()
+            if (sexFilterStatus != null) {
+                cattleList = CattleManager.filterListBySex(sexFilterStatus, cattleList)
+            }
+
+            // Check weight filter
+            val weightFilterStatus = preferenceUtils.getWeightFilter()
+            if (weightFilterStatus != null) {
+                cattleList = CattleManager.filterListByWeight(weightFilterStatus, cattleList)
+            }
+
+            // Order resulting list
+            val orderCriteria = preferenceUtils.getOrderBy()
+            // Reorder list
+            cattleList = CattleManager.orderListBy(orderCriteria, cattleList)
+
+            return cattleList
         }
     }
 }
