+++ app/src/main/java/com/iua/agustinpereyra/view/base/FilterableCattleRecyclerFragment.kt
@@ -40,32 +40,7 @@ abstract class FilterableCattleRecyclerFragment : Fragment(), SearchView.OnQuery
 
     override fun onResume() {
         super.onResume()
-        // TODO: Is it okay to put it here or should go at controller?
-        // Do only if already initiallized
-        // TODO: Is it right?
-        var cattleList = baseCattleList
-        val preferenceUtils = PreferenceUtils(context)
-
-        // Filtering
-        // Check sex filter
-        val sexFilterStatus = preferenceUtils.getSexFilter()
-        if (sexFilterStatus != null) {
-            cattleList = CattleManager.filterListBySex(sexFilterStatus, cattleList)
-        }
-
-        // Check weight filter
-        val weightFilterStatus = preferenceUtils.getWeightFilter()
-        if (weightFilterStatus != null) {
-            cattleList = CattleManager.filterListByWeight(weightFilterStatus, cattleList)
-        }
-
-        // Order resulting list
-        val orderCriteria = preferenceUtils.getOrderBy()
-        // Reorder list
-        cattleList = CattleManager.orderListBy(orderCriteria, cattleList)
-
-        // Swap to new adapter with updated data
-        //  TODO: Is this optimal? Maybe a method to insert and pop all data in Adapter is better
+        val cattleList = CattleManager.orderBasedOnPreferences(baseCattleList, context)
         currentCattleList = cattleList
         recyclerViewAdapter.setCattle(cattleList)
     }
